refactor(silverback-gatsby): extract gatsby sourcing to reusable plugin
