+++ apps/silverback-gatsby/package.json
@@ -7,6 +7,7 @@
   "dependencies": {
     "@-amazeelabs/silverback-drupal": "^1.2.26",
     "@amazeelabs/gatsby-theme-core": "^0.6.6",
+    "@amazeelabs/gatsby-source-silverback": "^1",
     "axios": "^0.21.1",
     "cypress": "^7.0.0",
     "dotenv": "^10.0.0",
