+++ packages/npm/@amazeelabs/gatsby-source-silverback/.gitignore
@@ -1,0 +1,9 @@
+gatsby-node.js
+helpers
+### MANAGED BY @amazeelabs/scaffold - START
+.eslintrc.js
+.jest-runner-eslintrc
+.lintstagedrc
+.prettierrc
+jest.config.js
+### MANAGED BY @amazeelabs/scaffold - END
