+++ packages/npm/@amazeelabs/gatsby-source-silverback/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "@amazeelabs/gatsby-source-silverback",
+  "version": "1.0.0",
+  "description": "Source data from a Silverback GraphQL schema.",
+  "main": "index.js",
+  "author": "Amazee Labs <development@amazeelabs.com>",
+  "license": "MIT",
+  "private": false,
+  "dependencies": {
+    "@amazeelabs/eslint-config": "^1.3.0",
+    "@amazeelabs/jest-preset": "^1.3.1",
+    "@amazeelabs/prettier-config": "^1.1.0",
+    "@amazeelabs/scaffold": "^1.2.2"
+  },
+  "devDependencies": {
+    "@amazeelabs/eslint-config": "^1.3.0",
+    "@amazeelabs/jest-preset": "^1.3.1",
+    "@amazeelabs/prettier-config": "^1.1.0",
+    "@types/jest": "^26.0.23",
+    "eslint": "^7.25.0",
+    "jest": "^26.6.3",
+    "lint-staged": "^10.5.4",
+    "typescript": "^4.2.4"
+  },
+  "scripts": {
+    "prepare": "amazee-scaffold && tsc",
+    "precommit": "lint-staged",
+    "test": "jest --passWithNoTests && if [ -f test.sh ]; then ./test.sh; fi",
+    "watch": "jest --watch"
+  }
+}
