+++ packages/npm/@amazeelabs/gatsby-source-silverback/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends":"@tsconfig/recommended/tsconfig.json",
+  "compilerOptions": {
+    "outDir": ".",
+    "rootDir": "src",
+  },
+  "exclude": ["node_modules"]
+}
